Move example app inside of a directory named "example"
When the React Native CLI's auto-linking command runs, it attempts to
load all native files from every dependency and add them as cocoapods
with the relevant podspec. This auto-linker is smart enough to ignore
example or test projects, but *only if* the examples are in a directory
named test, example, or sample[1].

If test projects are found in directories without those names, the RN
linker throws a warning that the project is missing a podspec. :(

react-native-sortable-listview doesn't actually have any native code, so
we don't need a podspec. Instead of adding one, we can put the sample
code in a named directory and squash that warning. :)

1: https://github.com/react-native-community/cli/blob/e703283d667fc6a5d64e1ac916a58460e75333b6/packages/platform-ios/src/config/findProject.ts#L32-L49